test: . move functions to helper file
