+++ tests/Renderer/RenderingTestHelpers.ts
@@ -2,13 +2,16 @@ import type { App } from 'obsidian';
 import { State } from '../../src/Obsidian/Cache';
 import type { FilterOrErrorMessage } from '../../src/Query/Filter/FilterOrErrorMessage';
 import { Query } from '../../src/Query/Query';
-import type { HTMLQueryRendererParameters } from '../../src/Renderer/HtmlQueryResultsRenderer';
+import type {
+    HTMLQueryRendererParameters,
+    HtmlQueryResultsRenderer,
+} from '../../src/Renderer/HtmlQueryResultsRenderer';
 import { MarkdownQueryResultsRenderer } from '../../src/Renderer/MarkdownQueryResultsRenderer';
 import type { Task } from '../../src/Task/Task';
 import { verifyWithFileExtension } from '../TestingTools/ApprovalTestHelpers';
 import { prettifyHTML } from '../TestingTools/HTMLHelpers';
-import { toMarkdown } from '../TestingTools/TestHelpers';
 import { createTestTasksFile } from '../TestingTools/TasksFileHelpers';
+import { toMarkdown } from '../TestingTools/TestHelpers';
 
 export const mockHTMLRenderer = async (_obsidianApp: App, text: string, element: HTMLSpanElement, _path: string) => {
     // Contrary to the default mockTextRenderer(),
@@ -70,7 +73,33 @@ ${toMarkdown(allTasks)}
     return { tasksAsMarkdown, prettyHTML };
 }
 
+export async function verifyHtmlFromRenderer(
+    renderer: HtmlQueryResultsRenderer,
+    state: State,
+    query: Query,
+    allTasks: Task[],
+) {
+    const container = document.createElement('div');
+    renderer.content = container;
+    await renderer.renderQuery(state, query.applyQueryToTasks(allTasks));
+
+    verifyRenderedTasks(container, allTasks);
+}
+
 export function verifyRenderedTasks(container: HTMLDivElement, allTasks: Task[]): void {
     const { tasksAsMarkdown, prettyHTML } = tasksMarkdownAndPrettifiedHtml(container, allTasks);
     verifyWithFileExtension(tasksAsMarkdown + prettyHTML, 'html');
+}
+
+export async function renderTasks(
+    state: State,
+    renderer: HtmlQueryResultsRenderer,
+    allTasks: Task[],
+    query: Query,
+): Promise<HTMLDivElement> {
+    const container = document.createElement('div');
+
+    renderer.content = container;
+    await renderer.renderQuery(state, query.applyQueryToTasks(allTasks));
+    return container;
 }
